# tests/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/circuits": {
      "get": {
        "operationId": "CircuitsController_listCircuits",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all circuits for the authenticated user.",
        "tags": [
          "Circuits"
        ]
      },
      "post": {
        "operationId": "CircuitsController_createCircuit",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create and persist a new circuit.",
        "tags": [
          "Circuits"
        ]
      }
    },
    "/api/v1/circuits/{id}": {
      "get": {
        "operationId": "CircuitsController_getCircuit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a circuit by ID (owned by the authenticated user).",
        "tags": [
          "Circuits"
        ]
      },
      "put": {
        "operationId": "CircuitsController_updateCircuit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a circuit's name and/or operations.",
        "tags": [
          "Circuits"
        ]
      },
      "delete": {
        "operationId": "CircuitsController_deleteCircuit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a circuit.",
        "tags": [
          "Circuits"
        ]
      }
    },
    "/api/v1/circuits/{id}/simulate": {
      "post": {
        "operationId": "CircuitsController_simulateCircuit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Simulate a circuit and return real results (statevector, probabilities\nand sampled measurement counts).\n\nThe circuit is resolved in one of two ways:\n - inline: the request body carries `numQubits` (+ `operations`), used as-is\n - stored: otherwise the persisted circuit `:id` is loaded and simulated",
        "tags": [
          "Circuits"
        ]
      }
    },
...
  },
  "info": {
    "title": "casimirQ API",
    "description": "REST API for the casimirQ quantum circuit simulation platform: authentication, circuits, simulation, asynchronous jobs, algorithms, visualizations, and advanced features (error correction, noise, quantum ML).",
    "version": "1.0.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Docker (nginx single-origin)"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local backend"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}